Rename pegasus package into dax
